bash: Refresh bash exports
Use asdf-vm as the main interface for language-specific package
managers. The `PATH` no longer needs to be modified for each language,
since `asdf` handles that with shims.

Use some known paths to set up `asdf`, `fzf`, etc. The paths used are
specific to Arch Linux, but could be found on other distributions. Mac
OS paths could also be added.

Add ~/.local/bin as a standard location for user binaries. This is
inline with the systemd file hierarchy[1], which is based off the XDG
Base Directory specification[2].

[1]: https://www.freedesktop.org/software/systemd/man/file-hierarchy.html
[2]: https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html